# swagger.yaml
...
swagger: '2.0'
info:
  title: "Siamo Veramente Euforici API"
  description: API description for Software Engineering II project
  version: '1.0'

#server
host: se2-project-2018.herokuapp.com
basePath: /v1
schemes:
- http
consumes:
- application/json
produces:
- application/json

#authentication token
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: X-API-Key
    description: >
      You must have an API-key in order to use all the APIs.

#global security (applies to all operations):
security:
  - ApiKeyAuth: []
paths:
...
  /classes:
    get:
      summary: List all the classes
      tags: [Class]
      responses:
        200:
          description: Successful response.
          schema:
            type: array
            items:
              $ref: '#/definitions/Class'
            description: It returns an array with all the classes.
        400:
          description: Ops! Something went wrong.
        404:
          description: We're sorry. No class found.

    post:
      summary: Create a new class
      tags: [Class]
      parameters:
        - in: body
          name: Class parameters
          description: The parameters necessary to create a new class (required fields for the definition "Class" are mandatory).
          schema:
            type: object
            properties:
              name:
                type: string
                description: The name of the new class.
              participants:
                type: array
                items:
                    $ref: '#/definitions/User'
                description: An array with the participants of the new class.
      responses:
        201:
          description: The class has been created.
          schema:
            $ref: '#/definitions/Class'
            description: It returns the created class.
        400:
          description: Ops! Something went wrong.

    delete:
      summary: Delete all classes
      tags: [Class]
      responses:
        204:
          description: All classes have been deleted.
        400:
          description: Ops! Something went wrong.
        404:
          description: We're sorry. No class found.
  /classes/{classId}:
    get:
      summary: Returns a class by ID
      tags: [Class]
      parameters:
        - in: path
          name: classId
          description: The ID necessary to return the specified class.
          required: true
          type: integer
          minimum: 1
      responses:
        200:
          description: Successful response.
          schema:
            $ref: '#/definitions/Class'
            description: It returns the selected class.
        400:
          description: Ops! Something went wrong.
        404:
          description: We're sorry. No class found with the given ID.

    put:
      summary: Updates a class by a given ID
      tags: [Class]
      parameters:
        - in: path
          name: classId
          description: The ID required to update the specified class.
          required: true
          type: integer
          minimum: 1
        - in: body
          name: The updated class.
          description: The class properties that can be changed.
          schema:
            $ref: '#/definitions/Class'
            description: >
              You can choose to modify only certain fields or the entire object.
              This is all handled by the API implementation.
              Of course, you can't change the ID of the object.
      responses:
        200:
          description: Class updated.
          schema:
            $ref: '#/definitions/Class'
            description: This response returns the updated class.
        400:
          description: Ops! Something went wrong.
        404:
          description: We're sorry. No class found with the given ID.
        409:
          description: Ops! It seems there are some conflicts.

    delete:
      summary: Delete the class with the given ID
      tags: [Class]
      parameters:
        - in: path
          name: classId
          description: The ID required to delete the specified class.
          required: true
          type: integer
          minimum: 1
      responses:
        204:
          description: Class deleted.
        400:
          description: Ops! Something went wrong.
        404:
          description: We're sorry. No class found with the given ID.
...
  /exams:
    get:
      summary: List all the exams
      tags: [Exam]
      responses:
        200:
          description: Successful response.
          schema:
            type: array
            items:
              $ref: '#/definitions/Exam'
            description: It returns an array with all the exams.
        400:
          description: Ops! Something went wrong.
        404:
          description: We're sorry. No exam found.

    post:
      summary: Create a new exam
      tags: [Exam]
      parameters:
        - in: body
          name: Exam parameters.
          description: The parameters necessary to create a new exam (required fields for the definition "exam" are mandatory).
          schema:
            type: object 
            properties:
              taskgroup:
                $ref: '#/definitions/TaskGroup'
                description: The set of all the tasks.
              startline:
                type: string
                description: When the exam starts.
              deadline:
                type: string
                description: When the exam expires.
              classes:
                type: array
                items:
                  $ref: '#/definitions/Class'
                description: The specific class/classes to which the exam is assigned.
              teacher:
                $ref: '#/definitions/User'
                description: The teacher who has assigned the exam to the class.
      responses:
        201:
          description: The exam has been created.
          schema:
            $ref: '#/definitions/Exam'
            description: It returns the created exam.
        400:
          description: Ops! Something went wrong.

    delete:
      summary: Delete all the exams
      tags: [Exam]
      responses:
        204:
          description: All exams have been deleted.
        400:
          description: Ops! Something went wrong.
        404:
          description: We're sorry. No exam found.
  /exams/{examId}:
    get:
      summary: Returns an exam by ID
      tags: [Exam]
      parameters:
        - in: path
          name: examId
          description: The ID required to return the specific exam.
          required: true
          type: integer
          minimum: 1
      responses:
        200:
          description: Successful response.
          schema:
            $ref: '#/definitions/Exam'
            description: It returns the selected exam.
        400:
          description: Ops! Something went wrong.
        404:
          description: We're sorry. No exam found with the given ID.

    put:  
      summary: Update a specific Exam with the given ID
      tags: [Exam]
      parameters:
        - in: path
          name: examId
          description: The ID required to return the specific exam.
          required: true
          type: integer
          minimum: 1
        - in: body
          name: The updated exam.
          description: The exam properties that can be changed.
          schema:
            $ref: '#/definitions/Exam'
            description: >
              You can choose to modify only certain fields or the entire object.
              This is all handled by the API implementation.
              Of course, you can't change the ID of the object.
      responses:
        200:
          description: Exam updated.
          schema:
            $ref: '#/definitions/Exam'
            description: This response returns the updated exam.
        400:
          description: Ops! Something went wrong.
        404:
          description: We're sorry. No exam found with the given ID.
        409:
          description: Ops! It seems there are some conflicts.

    delete:  
      summary: Delete an exam by a given ID
      tags: [Exam]
      parameters:
        - in: path
          name: examId
          description: The ID required to delete the specified exam.
          required: true
          type: integer
          minimum: 1
      responses:
        204:
          description: Exam deleted.
        400:
          description: Ops! Something went wrong.
        404:
          description: We're sorry. No exam found with the given ID.
...
definitions:
  User:
    title: user
    type: object
    properties:
      id:
        type: integer
        description: the user ID
        example: 22
      name:
        type: string
        description: the user's name
        example: 'Britney'
      surname:
        type: string
        description: the user's surname
        example: 'Spears'
      uniNumber:
        type: string
        description: the identifier for the user provided by the university
        example: '123456'
      isTeacher:
        type: boolean
        description: >
          this value is true if the user is a teacher, false otherwise
        example: false
      email:
        type: string
        description: the user's email
        example: 'britney@spears.com'
      password: # andrebbe criptata
        type: string
        description: the user's password
        example: 'abc123'
      examsList:
        type: array
        items:
          $ref: '#/definitions/Exam'
        description: the list of all the exams the user (if student) is taking
        example: [exam1, exam2, exam3]
    required:
    - id
    - name
    - surname
    - uniNumber
    - isTeacher
    - email
    - password
  Class:
    title: class
    type: object
    properties:
      id:
        type: integer
        description: the class ID
        example: 4
      name:
        type: string
        description: the class' name
        example: 'Siamo Veramente Euforici'
      participants:
        type: array
        items:
          $ref: '#/definitions/User'
        description: the collection of users taking part in a class
        example: [user1, user2, user3]
    required:
    - id
    - name
    - participants
  Task:
    title: task
    type: object
    properties:
      id:
        type: integer
        description: the task ID
        example: 5
      question:
        type: string
        description: the question
        example: 'Il nome di Rovazzi, il cantante'
      questionType:
        type: string
        enum: [multipleChoice, openAnswer]
        description: the type of question asked
        example: 'multipleChoice'
      choices:
        type: array
        items:
          type: string    
        description: if multipleChoice, all the possible answers to the question
        example: [Fabio, Luigi, Mario, Giovanni]
      answers:
        type: array
        items:
          type: string
        description: if multipleChoice, it describes all the correct answers
        example: [Fabio]
      teacher: 
        $ref: '#/definitions/User'
        description: the teacher who have created the task
        example: Pippo Baudo
    required:
    - id
    - question
    - questionType
    - teacher
  TaskGroup:
    title: taskgroup
    type: object
    properties: 
      id:
        type: integer
        description: the taskgroup ID
        example: 6
      name:
        type: string
        description: name given to the taskGroup
        example: 'Ingegneria del Software 2 - Questions pool #1'
      numberTasks:
        type: integer
        description: the number of questions selected for the task group
        example: 3
      tasks:
        type: array
        items:
          $ref: '#/definitions/Task'
        description: the array of all the tasks of the exam
        example: [task1, task2, task3]
    required:
    - id
    - tasks
  Exam:
    title: exam
    type: object
    properties:
      id: 
        type: integer
        description: the exam ID
        example: 7
      taskgroup: 
        $ref: '#/definitions/TaskGroup'
        description: the set of all the tasks
        example: Ingegneria del Software 2
      startline:
        type: string
        description: when the exam starts
        example: '19/12/2018 10:30'
      deadline:
        type: string
        description: the deadline for the delivery of the exam
        example: '19/12/2018 12:30'
      classes:
        type: array
        items:
          $ref: '#/definitions/Class'
        description: the specific class/classes to which the exam is assigned
        example: [class1, class2]
      teacher:
        $ref: '#/definitions/User'
        description: the teacher who has assigned the exam to the class
        example: Fabio Casati
    required:
    - id
    - taskgroup
    - deadline
    - classes
    - teacher
...